# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GetCarApi — vehicle data collection and VIN history API
servers:
  - url: /api
    description: Base API path
...
security: []   # No global security — applied per-endpoint below
paths:
...
  /admin/auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/auth/me:
    get:
      operationId: adminGetMe
      tags: [auth]
      summary: Get current admin user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: VDI token (vdi_…)
      description: >
        API token issued to your account. Pass as
        `Authorization: Bearer vdi_<token>`. Required for all VIN routes
        (including check) and live inventory endpoints.
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AdminLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 1
    AdminUser:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          type: number
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time

    # ── Dashboard ────────────────────────────────────────────────────────────
...